+++ server/handlers/changePlayerData.js
@@ -1,0 +1,10 @@
+function changePlayerData(io){
+    const socket = this;
+    const gameId = socket.request.session.gameId;
+    io.to(gameId).emit("receiveData")
+}
+
+
+module.exports = {
+    changePlayerData
+}