+++ server/index.js
@@ -1,15 +1,32 @@
-const express = require('express');
-const app = express();
-const socket = require('socket.io');
+const express = require('express')
+const app = express()
+const socket = require('socket.io')
+const session = require("express-session")
+const server = app.listen(3000)
+const io = socket(server)
+const { joinRoomHandler } = require("./handlers/joinRoomHandler")
+const { changePlayerData } = require("./handlers/changePlayerData")
+
+const sessionMiddleware = session({
+    secret: "WFXO1",
+    resave: false,
+    saveUninitialized: false
+})
 
 app.use(express.static("static"))
+app.use(sessionMiddleware)
 
-const server = app.listen(3000, () => {
-    console.log("listening on 3000")
-})
+global.rooms = []
 
-const io = socket(server);
+const connection = (socket) => {
+    socket.on("join", joinRoomHandler.bind(socket, io))
+    socket.on("changePlayerData", changePlayerData.bind(socket, io))
+}
 
-io.on('connection', () => {
-    console.log("Client connected")
-})
+// EXPRESS SESSION AVAILABLE IN SOCKET.IO
+// ACCESSIBLE VIA socket.request.session
+io.use((socket, next) => {
+    sessionMiddleware(socket.request, {}, next);
+});
+
+io.on('connection', connection)