+++ server/utils/obtainFreeRoom.js
@@ -1,0 +1,16 @@
+const { RoomsManager } = require("../components/RoomsManager")
+
+// @var object - room
+function obtainFreeRoom(){
+    const manager = new RoomsManager()
+    const freeRoom = manager.searchForFreeRoom();
+    if(freeRoom){
+        return freeRoom
+    }else{
+        return manager.createRoom();
+    }
+}
+
+module.exports = {
+    obtainFreeRoom: obtainFreeRoom
+}